requirements.txt
streamlit>=1.36
pydantic>=2.7
langchain>=0.2.12
langchain-community>=0.2.12
langchain-text-splitters>=0.2.2
langchain-openai>=0.1.22
langgraph>=0.2.26
faiss-cpu>=1.8
pypdf>=4.2
pymupdf>=1.24.1
python-dotenv>=1.0.1
langfuse>=2.39.0
numpy>=1.26
pandas>=2.2
tiktoken>=0.7
rich>=13.7
plotly>=5.22
wheel
httpx>=0.27
requests>=2.31